backend/requirements.txt
# --- Core Web Framework ---
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.42
Werkzeug==3.1.3

# --- Production Server ---
gunicorn==23.0.0
uvicorn==0.35.0

# --- AI & Vector DB ---
google-generativeai==0.8.5
pinecone-client==5.0.1

# --- Database Drivers ---
# Required for Render/Heroku Postgres
psycopg2-binary==2.9.10 

# --- Utilities ---
python-dotenv==1.1.1
python-dateutil==2.9.0.post0
requests==2.32.4
pydantic==2.11.7
pytz==2025.2
rich==14.1.0

# --- Low-level Deps (Kept for safety) ---
click==8.2.1
colorama==0.4.6
greenlet==3.2.4
httpx==0.28.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
typing_extensions==4.14.1
urllib3==2.5.0